Fix : 존재하지 않는 파일 삭제 요청
전체 데이터 삭제 요청 시 파일이 존재하지 않아도
파일 삭제를 요청하는 경우 발생하는 에러를 수정하였습니다.

+++ server/clinical-trial-system/src/main/java/com/example/clinicaltrialsystem/Storage/ImageFileStorageService.java
@@ -32,7 +32,12 @@ public class ImageFileStorageService {
     public File[] getAllFiles(){
         String dirPath=System.getProperty("user.home")+STORAGE_DIR_NAME;
         File dir=new File(dirPath);
-        return dir.listFiles();
+        if(dir.exists()){
+            return dir.listFiles();
+        }
+        else{
+            return null;
+        }
     }
 
     public boolean deleteFile(String filename){
@@ -42,7 +47,11 @@ public class ImageFileStorageService {
 
     public boolean deleteAll(){
         String path=System.getProperty("user.home")+STORAGE_DIR_NAME;
-        for(File file : getAllFiles()){
+        File[] subFileList=getAllFiles();
+        if(subFileList==null){
+            return true;
+        }
+        for(File file : subFileList){
             file.delete();
         }
         return new File(path).delete();
